+++ Days/Day02/Day02.cs
@@ -14,7 +14,7 @@ namespace AdventOfCode2021.Days.Day02
 
         [TestCase(Input.Example, 150)]
         [TestCase(Input.File, 1383564)]
-        public override int Part1(List<Day02Data> input)
+        public override long Part1(List<Day02Data> input)
         {
             var depth = input.Where(it => it.Instruction == Day02Enum.Down).Sum(it => it.Magnitude) -
                         input.Where(it => it.Instruction == Day02Enum.Up).Sum(it => it.Magnitude);
@@ -24,9 +24,9 @@ namespace AdventOfCode2021.Days.Day02
 
         [TestCase(Input.Example, 900)]
         [TestCase(Input.File, 1488311643)]
-        public override int Part2(List<Day02Data> input)
+        public override long Part2(List<Day02Data> input)
         {
-            var result = input.Aggregate(new { Depth = 0, Position = 0, Aim = 0 }, (accum, current) =>
+            var result = input.Aggregate(new { Depth = 0l, Position = 0l, Aim = 0l }, (accum, current) =>
                 current.Instruction switch
                 {
                     Day02Enum.Down => new { accum.Depth, accum.Position, Aim = accum.Aim + current.Magnitude },
@@ -42,7 +42,7 @@ namespace AdventOfCode2021.Days.Day02
     public class Day02Data
     {
         public Day02Enum Instruction { get; set; }
-        public int Magnitude { get; set; }
+        public long Magnitude { get; set; }
     }
 
     public enum Day02Enum
