+++ Days/Day05/Day05.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using AdventOfCode2015.Utils;
 using AdventOfCode2021.Utils;
 using JetBrains.Annotations;
 
@@ -25,7 +24,7 @@ namespace AdventOfCode2021.Days.Day05
 
         [TestCase(Input.Example, 5)]
         [TestCase(Input.File, 6225)]
-        public override int Part1(List<Day05Input> input)
+        public override long Part1(List<Day05Input> input)
         {
             bool IsOrthogonal(Day05Input it) => it.X1 == it.X2 || it.Y1 == it.Y2;
             return Part2(input.Where(IsOrthogonal).ToList());
@@ -33,9 +32,9 @@ namespace AdventOfCode2021.Days.Day05
 
         [TestCase(Input.Example, 12)]
         [TestCase(Input.File, 22116)]
-        public override int Part2(List<Day05Input> input)
+        public override long Part2(List<Day05Input> input)
         {
-            var grid = new Dictionary<Position, int>();
+            var grid = new Dictionary<Position, long>();
             foreach (var line in input)
             {
                 var dX = MoreMath.Sign(line.X2 - line.X1);
@@ -57,12 +56,12 @@ namespace AdventOfCode2021.Days.Day05
     public class Day05Input
     {
         [RxFormat(After = ",")]
-        public int X1 { get; set; }
+        public long X1 { get; set; }
         [RxFormat(After = "->")]
-        public int Y1 { get; set; }
+        public long Y1 { get; set; }
 
         [RxFormat(After = ",")]
-        public int X2 { get; set; }
-        public int Y2 { get; set; }
+        public long X2 { get; set; }
+        public long Y2 { get; set; }
     }
 }
