+++ Utils/MoreMath.cs
@@ -2,12 +2,12 @@
 {
     public static class MoreMath
     {
-        public static int Sign(int a) => a switch
+        public static long Sign(long a) => a switch
         {
             < 0 => -1,
             > 0 => 1,
             _ => 0
         };
-        public static int Abs(int a) => a < 0 ? -a : a;
+        public static long Abs(long a) => a < 0 ? -a : a;
     }
 }
